Using LSTM for prediction of future census data of each tract (#3, #5)
The model is trained with x being data from year 2008 to 2017, y being data from year 2018. The trained model is them used recursively to predict future values, up till 2028, based on data from year 2009 to 2018.

When training, 25% of the data is used as the testing set, the model achieved accuracy of 95%.

NB model and output csv files are tracked using git lfs, please ensure it is installed

Former-commit-id: c0904803531d5d6c6cf102cc79ffc44abef33ba4 [formerly 8d0d973d72da3dfd8ab9c1f934291e8a3161a187]
Former-commit-id: cb2c8ac66a29348c665f4a27c0b2acd1a3e8fcb3